scripts/requirements.txt
PyGithub==1.35
pyyaml==3.12